Added cryptographically secure tripcodes.  This system uses the SHA-512 HMAC from OpenSSL.
The way HMACs work is they take a key and a message, and then hash them such that the resultant hash is immune to length extension attacks.  In this case, the key is the password part (the bit after the # symbol in the name field) and the message is the name part (the part before the # symbol).

Current tripcode technology is flawed because it is very easy for someone to find a collision and pretent to be you.  The SHA-512 hash algorithm is considered cryptographically secure, so this is not a problem.  The only problem is because this tripcode system doesn't work like the old ones, tripcodes generated here will be different from tripcodes generated on other chans using the common tripcode algorithm.  It's a tradeoff for security, but a worthiwhile one.